move all relevant types for DataResolver interface to connector project
