+++ server/connectors/api-connector-mysql/src/main/scala/com/prisma/api/connector/mysql/database/QueryArguments.scala
@@ -1,6 +1,5 @@
 package com.prisma.api.connector.mysql.database
 
-import com.prisma.api.connector.mysql.database.Types.DataItemFilterCollection
 import com.prisma.api.schema.APIErrors
 import com.prisma.shared.models.{Field, TypeIdentifier}
 import slick.jdbc.SQLActionBuilder
