+++ server/connectors/api-connector-mysql/src/main/scala/com/prisma/api/connector/mysql/database/Types.scala
@@ -1,12 +1,7 @@
 package com.prisma.api.connector.mysql.database
 
-import com.prisma.api.connector.mysql.database.Types.DataItemFilterCollection
+import com.prisma.api.connector.Types.DataItemFilterCollection
 import com.prisma.shared.models.{Field, Model, Relation}
-
-object Types {
-  type DataItemFilterCollection = Seq[_ >: Seq[Any] <: Any]
-//  type UserData                 = Map[String, Option[Any]]
-}
 
 case class FilterElement(key: String,
                          value: Any,
@@ -15,5 +10,3 @@ case class FilterElement(key: String,
                          relatedFilterElement: Option[FilterElementRelation] = None)
 
 case class FilterElementRelation(fromModel: Model, toModel: Model, relation: Relation, filter: DataItemFilterCollection)
-
-case class ScalarListValue(nodeId: String, position: Int, value: Any)
