+++ server/integration-tests/integration-tests-mysql/src/test/scala/com/prisma/integration/IntegrationBaseSpec.scala
@@ -2,7 +2,8 @@ package com.prisma.integration
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
-import com.prisma.api.connector.mysql.database.{DataResolver, DataResolverImpl}
+import com.prisma.api.connector.DataResolver
+import com.prisma.api.connector.mysql.database.DataResolverImpl
 import com.prisma.api.util.StringMatchers
 import com.prisma.api.{ApiDependenciesForTest, ApiTestDatabase, ApiTestServer}
 import com.prisma.deploy.specutils.{DeployTestDependencies, DeployTestServer}
