+++ server/servers/api/src/main/scala/com/prisma/api/ApiDependencies.scala
@@ -2,8 +2,8 @@ package com.prisma.api
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
-import com.prisma.api.connector.mysql.database.{DataResolver, DataResolverImpl, Databases}
-import com.prisma.api.connector.{ApiConnector, DatabaseMutactionExecutor}
+import com.prisma.api.connector.mysql.database.{DataResolverImpl, Databases}
+import com.prisma.api.connector.{ApiConnector, DataResolver, DatabaseMutactionExecutor}
 import com.prisma.api.mutactions.{DatabaseMutactionVerifier, SideEffectMutactionExecutor}
 import com.prisma.api.project.ProjectFetcher
 import com.prisma.api.resolver.DeferredResolverProvider
