+++ server/servers/api/src/main/scala/com/prisma/api/mutations/ClientMutation.scala
@@ -1,7 +1,6 @@
 package com.prisma.api.mutations
 
-import com.prisma.api.connector.mysql.database.DataResolver
-import com.prisma.api.connector.{DataItem, DatabaseMutaction, NodeSelector, SideEffectMutaction}
+import com.prisma.api.connector._
 import com.prisma.shared.models.IdType.Id
 import cool.graph.cuid.Cuid
