+++ server/servers/api/src/main/scala/com/prisma/api/mutations/Create.scala
@@ -3,8 +3,7 @@ package com.prisma.api.mutations
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import com.prisma.api.ApiDependencies
-import com.prisma.api.connector.mysql.database.DataResolver
-import com.prisma.api.connector.{CoolArgs, NodeSelector}
+import com.prisma.api.connector.{CoolArgs, DataResolver, NodeSelector}
 import com.prisma.api.mutactions.{DatabaseMutactions, ServerSideSubscriptions, SubscriptionEvents}
 import com.prisma.shared.models.IdType.Id
 import com.prisma.shared.models._
