+++ server/servers/api/src/main/scala/com/prisma/api/mutations/Delete.scala
@@ -3,8 +3,7 @@ package com.prisma.api.mutations
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import com.prisma.api.ApiDependencies
-import com.prisma.api.connector.mysql.database.DataResolver
-import com.prisma.api.connector.{CoolArgs, DataItem, NodeSelector, Path}
+import com.prisma.api.connector._
 import com.prisma.api.mutactions.{DatabaseMutactions, ServerSideSubscriptions, SubscriptionEvents}
 import com.prisma.api.schema.{APIErrors, ObjectTypeBuilder}
 import com.prisma.shared.models.IdType.Id
