+++ server/servers/api/src/main/scala/com/prisma/api/mutations/UpdateMany.scala
@@ -1,9 +1,8 @@
 package com.prisma.api.mutations
 
 import com.prisma.api.ApiDependencies
-import com.prisma.api.connector.mysql.database.DataResolver
-import com.prisma.api.connector.mysql.database.Types.DataItemFilterCollection
-import com.prisma.api.connector.{CoolArgs, UpdateDataItems}
+import com.prisma.api.connector.Types.DataItemFilterCollection
+import com.prisma.api.connector.{CoolArgs, DataResolver, UpdateDataItems}
 import com.prisma.shared.models.{Model, Project}
 import sangria.schema
