+++ server/servers/api/src/main/scala/com/prisma/api/resolver/CountManyModelDeferredResolver.scala
@@ -1,6 +1,6 @@
 package com.prisma.api.resolver
 
-import com.prisma.api.connector.mysql.database.DataResolver
+import com.prisma.api.connector.DataResolver
 import com.prisma.api.resolver.DeferredTypes.{CountManyModelDeferred, OrderedDeferred, OrderedDeferredFutureResult}
 
 class CountManyModelDeferredResolver(dataResolver: DataResolver) {
