+++ server/servers/api/src/main/scala/com/prisma/api/resolver/ManyModelDeferredResolver.scala
@@ -1,5 +1,6 @@
 package com.prisma.api.resolver
 
+import com.prisma.api.connector.{DataResolver, ResolverResult}
 import com.prisma.api.connector.mysql.database._
 import com.prisma.api.resolver.DeferredTypes._
