+++ server/servers/api/src/main/scala/com/prisma/api/schema/ObjectTypeBuilder.scala
@@ -1,8 +1,8 @@
 package com.prisma.api.schema
 
-import com.prisma.api.connector.{DataItem, OrderBy, QueryArguments}
-import com.prisma.api.connector.mysql.database.Types.DataItemFilterCollection
+import com.prisma.api.connector.Types.DataItemFilterCollection
 import com.prisma.api.connector.mysql.database._
+import com.prisma.api.connector.{DataItem, OrderBy, QueryArguments, Types}
 import com.prisma.api.mutations.BatchPayload
 import com.prisma.api.resolver.DeferredTypes._
 import com.prisma.api.resolver.{IdBasedConnection, IdBasedConnectionDefinition}
